feat : implement a v1 version of gameModel (issue #9)
The model contains several hints. For this I created a hint class which represents them. It currently makes a dummy call to an api and doesn't use parsing this will be changed as soon as fetching to the api and parsing are implemented.
This model contains many getters and a getNewHint() method which reveals a new hint randomly.

To keep a version that compiles, I had to change example.vue and infoboxView.vue.