feat : added a HintList class representing all the hints (#9) this implementation is in HintList.ts, the definition of the Hint<T> class has also moved there for more coherence. It had to be a class and not an interface in order to have the toList() function which is very useful when filtering between hints is needed. Now to acces info about the myterious celebrity you get it by model.hints.birth instead of model.birth before. GameModel now has less attributes and has a HintList. I modified infoboxView.vue to make it compatible with this new version.
Moreover, I changed the constructor of a hint so that you can construct them without giving the value, only the level of the hint is needed. This is useful so that HintList class can initiate all its hint without a constructor, otherwise it would have a needed a constructor with a bunch of parameters (birth, death, occupation, citizenship, etc.) and it would have made it a hassle to create a new HintList. HintLists can now be instanciated without any parameters.